Extracted genres formatting
